+++ src/dbt_score/__init__.py
@@ -1,17 +1,18 @@
 """Init dbt_score package."""
 
-from dbt_score.models import Model, Snapshot, Source
+from dbt_score.models import Exposure, Model, Snapshot, Source
 from dbt_score.rule import Rule, RuleViolation, Severity, rule
 from dbt_score.rule_filter import RuleFilter, rule_filter
 
 __all__ = [
+    "Exposure",
     "Model",
-    "Source",
-    "Snapshot",
-    "RuleFilter",
     "Rule",
+    "RuleFilter",
     "RuleViolation",
     "Severity",
-    "rule_filter",
+    "Snapshot",
+    "Source",
     "rule",
+    "rule_filter",
 ]