+++ src/dbt_score/dbt_utils.py
@@ -69,7 +69,15 @@ def dbt_parse() -> "dbtRunnerResult":
 @dbt_required
 def dbt_ls(select: Iterable[str] | None) -> Iterable[str]:
     """Run dbt ls."""
-    cmd = ["ls", "--resource-types", "model", "source", "snapshot", "--output", "name"]
+    cmd = [
+        "ls",
+        "--resource-types",
+        "model",
+        "source",
+        "exposure",
+        "--output",
+        "name",
+    ]
     if select:
         cmd += ["--select", *select]