+++ tests/test_rule.py
@@ -1,7 +1,16 @@
 """Test rule."""
 
 import pytest
-from dbt_score import Model, Rule, RuleViolation, Severity, Snapshot, Source, rule
+from dbt_score import (
+    Exposure,
+    Model,
+    Rule,
+    RuleViolation,
+    Severity,
+    Snapshot,
+    Source,
+    rule,
+)
 from dbt_score.rule_filter import RuleFilter, rule_filter
 
 
@@ -80,6 +89,10 @@ def test_missing_evaluate_rule_class(model1):
         ("decorator_rule_no_parens_snapshot", Snapshot),
         ("decorator_rule_args_snapshot", Snapshot),
         ("class_rule_snapshot", Snapshot),
+        ("decorator_rule_exposure", Exposure),
+        ("decorator_rule_no_parens_exposure", Exposure),
+        ("decorator_rule_args_exposure", Exposure),
+        ("class_rule_exposure", Exposure),
     ],
 )
 def test_rule_introspects_its_resource_type(request, rule_fixture, expected_type):